Removed necessity to download the test graph

+++ MR-OCP/MROCPdjango/computation/composite/test_inv.sh
@@ -24,7 +24,11 @@ set -e
 echo "Testing invariants ..."
 mkdir -p ./mrdata/graphs
 GRAPH_FN="./mrdata/graphs/test.graphml"
-wget -O $GRAPH_FN "http://openconnecto.me/data/public/MR-data/Graphs/test.graphml"
+
+if  [ ! -f $GRAPH_FN ]; then
+  echo "Downloading the graph"
+  wget -O $GRAPH_FN "http://openconnecto.me/data/public/MR-data/Graphs/test.graphml"
+fi
 
 ./inv_exec -h
 echo "Testing All ..."
